서버에서 로컬 파일 정리하는 코드 클래스로 추출
